Update ATM2
The user is asked to enter their bank balance. After that, a menu is displayed with available options:

Check Balance: Displays the user's current balance.

Withdrawal: The user can withdraw an amount, provided it is less than the balance and no more than 1000 euros. Otherwise, an error message is displayed.

Deposit: The user enters an amount to deposit, and the balance is updated accordingly.