mygrep
This script accepts the path of a directory as an argument. After checking for validity (ensuring a single valid argument is passed), the program pauses for 2 seconds using the sleep command. It then initializes three counters, which are updated based on the contents of the directory. The results are displayed with appropriate messages.