sizecount
By executing the script ./sizecount <time>, where time is an integer between 1 and 24, the user searches for files created during a specific hour (from hour:00:00 to hour:59:59). Non-integer values or numbers outside this range are rejected. The find command is used to search for files, and any found are saved to the timefiles file.