telcat
When the user types ./telcat in the terminal, a selection menu appears. This comes from the first piece of code. Depending on the user's selection, the case statement from the second part of the code sends the user to the appropriate function.

User Options:
Add New Contact:
Using the command ./telcat -a, the user is asked to input details for a new contact, such as first name, last name, city, and phone number. Once entered, the data is saved to the file katalogos.

Display All Saved Contacts:
The command ./telcat -l shows all saved contacts on the user's console using the cat command. If the list is empty, a corresponding message is displayed.

Sort Contacts:
The command ./telcat -s <number> allows the user to sort contacts based on the following:

1: First name
2: Last name
3: City
4: Phone number
Each number corresponds to a sorting option.
Search for Contact:
Using ./telcat -c <keyword>, the user can search for contacts. If no matching entry is found, or the file is empty, an appropriate error message is displayed.

Delete a Contact:
The command ./telcat -d <keyword> allows the user to delete contacts from the list. If no matching keyword is found or the file is empty, an error message is shown.

Show Number of Contacts:
The command ./telcat -n shows the total number of contacts. If there are no entries, a suitable message is displayed.